migration issue fixed

+++ db/migrate/20200529043834_remove_column_from_homepage_sections.rb
@@ -1,5 +1,7 @@
 class RemoveColumnFromHomepageSections < ActiveRecord::Migration[5.1]
   def change
-  	remove_column :homepage_sections, :testimonial_main_heading_new, :text
+    if column_exists? :homepage_sections, :testimonial_main_heading_new
+      remove_column :homepage_sections, :testimonial_main_heading_new, :text
+    end
   end
 end
